+++ x/auth/ante/expected_keepers.go
@@ -1,0 +1,16 @@
+package ante
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/auth/exported"
+	"github.com/cosmos/cosmos-sdk/x/auth/types"
+)
+
+// AccountKeeper defines the contract needed for AccountKeeper related APIs.
+// Interface provides support to use non-sdk AccountKeeper for AnteHandler's decorators.
+type AccountKeeper interface {
+	GetParams(ctx sdk.Context) (params types.Params)
+	GetAccount(ctx sdk.Context, addr sdk.AccAddress) exported.Account
+	SetAccount(ctx sdk.Context, acc exported.Account)
+	GetModuleAddress(moduleName string) sdk.AccAddress
+}