+++ x/bank/bench_test.go
@@ -11,10 +11,9 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/auth"
 	authexported "github.com/cosmos/cosmos-sdk/x/auth/exported"
 	"github.com/cosmos/cosmos-sdk/x/staking"
-	"github.com/cosmos/cosmos-sdk/x/supply"
 )
 
-var moduleAccAddr = supply.NewModuleAddress(staking.BondedPoolName)
+var moduleAccAddr = auth.NewModuleAddress(staking.BondedPoolName)
 
 func BenchmarkOneBankSendTxPerBlock(b *testing.B) {
 	// Add an account at genesis