+++ x/bank/types/expected_keepers.go
@@ -8,6 +8,7 @@ import (
 // AccountKeeper defines the account contract that must be fulfilled when
 // creating a x/bank keeper.
 type AccountKeeper interface {
+	NewAccount(sdk.Context, exported.Account) exported.Account
 	NewAccountWithAddress(ctx sdk.Context, addr sdk.AccAddress) exported.Account
 
 	GetAccount(ctx sdk.Context, addr sdk.AccAddress) exported.Account
@@ -15,4 +16,12 @@ type AccountKeeper interface {
 	SetAccount(ctx sdk.Context, acc exported.Account)
 
 	IterateAccounts(ctx sdk.Context, process func(exported.Account) bool)
+
+	ValidatePermissions(macc exported.ModuleAccountI) error
+
+	GetModuleAddress(moduleName string) sdk.AccAddress
+	GetModuleAddressAndPermissions(moduleName string) (addr sdk.AccAddress, permissions []string)
+	GetModuleAccountAndPermissions(ctx sdk.Context, moduleName string) (exported.ModuleAccountI, []string)
+	GetModuleAccount(ctx sdk.Context, moduleName string) exported.ModuleAccountI
+	SetModuleAccount(ctx sdk.Context, macc exported.ModuleAccountI)
 }