+++ x/distribution/module_test.go
@@ -3,13 +3,12 @@ package distribution_test
 import (
 	"testing"
 
-	"github.com/cosmos/cosmos-sdk/x/distribution"
-
 	"github.com/stretchr/testify/require"
 	"github.com/tendermint/tendermint/abci/types"
 
 	"github.com/cosmos/cosmos-sdk/simapp"
-	"github.com/cosmos/cosmos-sdk/x/supply"
+	"github.com/cosmos/cosmos-sdk/x/auth"
+	"github.com/cosmos/cosmos-sdk/x/distribution"
 )
 
 func TestItCreatesModuleAccountOnInitBlock(t *testing.T) {
@@ -23,6 +22,6 @@ func TestItCreatesModuleAccountOnInitBlock(t *testing.T) {
 		},
 	)
 
-	acc := app.AccountKeeper.GetAccount(ctx, supply.NewModuleAddress(distribution.ModuleName))
+	acc := app.AccountKeeper.GetAccount(ctx, auth.NewModuleAddress(distribution.ModuleName))
 	require.NotNil(t, acc)
 }