+++ x/ibc/20-transfer/alias.go
@@ -49,7 +49,6 @@ type (
 	ChannelKeeper                      = types.ChannelKeeper
 	ClientKeeper                       = types.ClientKeeper
 	ConnectionKeeper                   = types.ConnectionKeeper
-	SupplyKeeper                       = types.SupplyKeeper
 	FungibleTokenPacketData            = types.FungibleTokenPacketData
 	FungibleTokenPacketAcknowledgement = types.FungibleTokenPacketAcknowledgement
 	MsgTransfer                        = types.MsgTransfer